requirements.txt
anyio==3.6.2
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
fastapi==0.91.0
flake8==6.0.0
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
mccabe==0.7.0
git+https://gitlab.crpo.su/vgds/asyncmonolog.git
git+https://git@gitlab.crpo.su/pyserv/akonfig.git
motor==3.1.1
pycodestyle==2.10.0
pydantic==1.10.4
pyflakes==3.0.1
pymongo==4.3.3
python-dotenv==0.21.1
python-multipart==0.0.5
python-telegram-bot==20.1
redis==4.5.4
requests==2.28.2
rfc3986==1.5.0
rocketchat-API==1.29.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.24.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
